+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added|Changed|Deprecated|Removed|Fixed|Security
 Nothing so far
 
+## 2.12.1 - 2020-01-08
+### Fixed
+- Linter fixes.
+
+## 2.12.0 - 2020-01-08
+### Added
+- Added `cache` and `string` mapping helper.
+- Added `before` and `after` filter helpers.
+
 ## 2.11.0 - 2019-04-29
 ### Added
 - Added `CollapseIterator` usable from php or twig.